+++ tests/testthat/test-ggbetweenstats.R
@@ -44,7 +44,7 @@ testthat::test_that(
     testthat::expect_equal(dim(pb$data[[1]]), c(44L, 13L))
     testthat::expect_equal(dim(pb$data[[2]]), c(4L, 26L))
     testthat::expect_equal(dim(pb$data[[3]]), c(2048L, 21L))
-    testthat::expect_equal(dim(pb$data[[5]]), c(4L, 13L))
+    testthat::expect_equal(dim(pb$data[[5]]), c(4L, 10L))
 
     # data from difference layers
     testthat::expect_equal(length(pb$data), 6L)
@@ -107,6 +107,17 @@ testthat::test_that(
     testthat::expect_equal(ggplot2::layer_scales(p)$y$range$range,
       c(0.00014, 5.71200000),
       tolerance = 1e-5
+    )
+
+    # checking centrality
+    testthat::expect_equal(
+      pb$data[[6]]$label,
+      c(
+        "list(~italic(widehat(mu))[mean]=='0.07926')",
+        "list(~italic(widehat(mu))[mean]=='0.62160')",
+        "list(~italic(widehat(mu))[mean]=='0.02155')",
+        "list(~italic(widehat(mu))[mean]=='0.14573')"
+      )
     )
 
     # checking x-axis sample size labels
@@ -176,7 +187,6 @@ testthat::test_that(
         x = "cyl",
         y = "n",
         type = "np",
-        mean.ci = TRUE,
         k = 2L,
         pairwise.comparisons = FALSE,
         conf.level = 0.90,
@@ -219,9 +229,9 @@ testthat::test_that(
     testthat::expect_identical(
       pb$data[[6]]$label,
       c(
-        "list(~italic(widehat(mu))=='2.29')",
-        "list(~italic(widehat(mu))=='3.12')",
-        "list(~italic(widehat(mu))=='4.00')"
+        "list(~italic(widehat(mu))[median]=='2.20')",
+        "list(~italic(widehat(mu))[median]=='3.21')",
+        "list(~italic(widehat(mu))[median]=='3.76')"
       )
     )
 
@@ -246,8 +256,8 @@ testthat::test_that(
     testthat::expect_identical(
       pb1$data[[6]]$label,
       c(
-        "list(~italic(widehat(mu))=='0.98')",
-        "list(~italic(widehat(mu))=='1.39')"
+        "list(~italic(widehat(mu))[mean]=='0.98')",
+        "list(~italic(widehat(mu))[mean]=='1.39')"
       )
     )
   }
@@ -285,6 +295,7 @@ testthat::test_that(
         data = ToothGrowth,
         x = supp,
         y = len,
+        type = "r",
         results.subtitle = FALSE,
         effsize.noncentral = FALSE,
         plot.type = "violin",
@@ -292,7 +303,6 @@ testthat::test_that(
         outlier.coef = 0.75,
         outlier.color = "blue",
         bf.message = FALSE,
-        mean.plotting = FALSE,
         sample.size.label = FALSE,
         package = "wesanderson",
         palette = "Royal1"
@@ -307,9 +317,6 @@ testthat::test_that(
     testthat::expect_null(pb1$plot$labels$subtitle, NULL)
     testthat::expect_null(pb1$plot$labels$caption, NULL)
     testthat::expect_null(pb2$plot$labels$caption, NULL)
-    testthat::expect_identical(length(pb1$data), 5L)
-    testthat::expect_identical(length(pb1$data), 5L)
-    testthat::expect_identical(length(pb2$data), 4L)
     testthat::expect_identical(
       unique(pb1$data[[1]]$colour),
       c("#D95F02FF", "#1B9E77FF")
@@ -322,9 +329,9 @@ testthat::test_that(
     # dims for data
     testthat::expect_equal(dim(pb1$data[[1]]), c(58L, 13L))
     testthat::expect_equal(dim(pb1$data[[2]]), c(2L, 26L))
-    testthat::expect_equal(dim(pb1$data[[4]]), c(2L, 13L))
+    testthat::expect_equal(dim(pb1$data[[4]]), c(2L, 10L))
     testthat::expect_equal(dim(pb1$data[[2]]), c(2L, 26L))
-    testthat::expect_equal(dim(pb1$data[[4]]), c(2L, 13L))
+    testthat::expect_equal(dim(pb1$data[[4]]), c(2L, 10L))
     testthat::expect_equal(dim(pb2$data[[1]]), c(58L, 13L))
     testthat::expect_equal(dim(pb2$data[[2]]), c(2L, 10L))
     testthat::expect_equal(dim(pb2$data[[3]]), c(1024L, 21L))
@@ -434,30 +441,21 @@ testthat::test_that(
             "mapped_discrete",
             "numeric"
           )),
-          group = 1:2,
-          y = c(
-            20.6633333333333,
-            16.9633333333333
-          ),
-          ymin = c(NA_real_, NA_real_),
-          ymax = c(
-            NA_real_,
-            NA_real_
-          ),
-          PANEL = structure(c(1L, 1L), .Label = "1", class = "factor"),
-          flipped_aes = c(FALSE, FALSE),
+          y = c(24.9087976539589, 15.4354838709677),
+          PANEL = structure(c(
+            1L,
+            1L
+          ), .Label = "1", class = "factor"),
+          group = structure(1:2, n = 2L),
           shape = c(19, 19),
-          colour = c(
-            "darkgreen",
-            "darkgreen"
-          ),
+          colour = c("darkgreen", "darkgreen"),
           size = c(5, 5),
           fill = c(NA, NA),
-          alpha = c(
-            NA,
-            NA
-          ),
-          stroke = c(0.5, 0.5)
+          alpha = c(NA, NA),
+          stroke = c(
+            0.5,
+            0.5
+          )
         ),
         row.names = c(NA, -2L),
         class = "data.frame"
@@ -472,10 +470,10 @@ testthat::test_that(
             "mapped_discrete",
             "numeric"
           )),
-          y = c(20.6633333333333, 16.9633333333333),
+          y = c(24.9087976539589, 15.4354838709677),
           label = c(
-            "list(~italic(widehat(mu))=='20.66')",
-            "list(~italic(widehat(mu))=='16.96')"
+            "list(~italic(widehat(mu))[MAP]=='24.91')",
+            "list(~italic(widehat(mu))[MAP]=='15.44')"
           ),
           PANEL = structure(c(
             1L,
@@ -486,39 +484,45 @@ testthat::test_that(
             "mapped_discrete",
             "numeric"
           )),
-          nudge_y = c(20.6633333333333, 16.9633333333333),
+          nudge_y = c(24.9087976539589, 15.4354838709677),
           colour = c("blue", "blue"),
           fill = c("white", "white"),
-          size = c(
-            3.88,
-            3.88
-          ),
+          size = c(3.88, 3.88),
           angle = c(0, 0),
           alpha = c(NA, NA),
-          family = c(
-            "",
-            ""
-          ),
+          family = c("", ""),
           fontface = c(1, 1),
-          lineheight = c(1.2, 1.2),
-          hjust = c(
-            0.5,
-            0.5
+          lineheight = c(
+            1.2,
+            1.2
           ),
+          hjust = c(0.5, 0.5),
           vjust = c(0.5, 0.5),
-          point.size = c(1, 1),
+          point.size = c(
+            1,
+            1
+          ),
           segment.linetype = c(4, 4),
           segment.size = c(0.5, 0.5),
           segment.curvature = c(0, 0),
           segment.angle = c(90, 90),
           segment.ncp = c(1, 1),
           segment.shape = c(0.5, 0.5),
-          segment.square = c(TRUE, TRUE),
+          segment.square = c(
+            TRUE,
+            TRUE
+          ),
           segment.squareShape = c(1, 1),
-          segment.inflect = c(FALSE, FALSE),
+          segment.inflect = c(
+            FALSE,
+            FALSE
+          ),
           segment.debug = c(FALSE, FALSE)
         ),
-        row.names = c(NA, -2L),
+        row.names = c(
+          NA,
+          -2L
+        ),
         class = "data.frame"
       )
     )
@@ -579,6 +583,14 @@ testthat::test_that(
         ),
         row.names = c(NA, -2L),
         class = "data.frame"
+      )
+    )
+
+    testthat::expect_equal(
+      pb2$data[[6]]$label,
+      c(
+        "list(~italic(widehat(mu))[trimmed]=='21.04')",
+        "list(~italic(widehat(mu))[trimmed]=='16.58')"
       )
     )
   }